+++ fabrictestbed_extensions/fablib/interface.py
@@ -22,6 +22,11 @@
 # SOFTWARE.
 #
 # Author: Paul Ruth (pruth@renci.org)
+
+"""
+Methods to work with FABRIC network interfaces.
+"""
+
 from __future__ import annotations
 
 import ipaddress
@@ -54,11 +59,15 @@ class Interface:
 
     def __init__(self, component: Component = None, fim_interface: FimInterface = None):
         """
-        Constructor. Sets keyword arguments as instance fields.
+        .. note::
+
+            Objects of this class are not created directly.
 
         :param component: the component to set on this interface
         :type component: Component
-        :param fim_interface: the FABRIC information model interface to set on this fablib interface
+
+        :param fim_interface: the FABRIC information model interface
+            to set on this fablib interface
         :type fim_interface: FimInterface
         """
         super().__init__()