+++ fabrictestbed_extensions/fablib/network_service.py
@@ -22,6 +22,13 @@
 # SOFTWARE.
 #
 # Author: Paul Ruth (pruth@renci.org)
+
+"""
+Methods to work with FABRIC `network services`_.
+
+.. _`network services`: https://learn.fabric-testbed.net/knowledge-base/glossary/#network_service
+"""
+
 from __future__ import annotations
 
 import logging
@@ -566,14 +573,16 @@ class NetworkService:
         self, slice: Slice = None, fim_network_service: FimNetworkService = None
     ):
         """
-        Not inteded for API use.
+        .. note::
 
-        Constructor. Sets the fablib slice and the FABRIC network service.
+            Not inteded for API use.
 
         :param slice: the fablib slice to set as instance state
         :type slice: Slice
-        :param fim_network_service: the FIM network service to set as instance state
-        :type fim_network_service: FIMNetworkService
+
+        :param fim_network_service: the FIM network service to set as
+            instance state
+        :type fim_network_service: FimNetworkService
         """
         super().__init__()
         self.fim_network_service = fim_network_service